make-category auth로 옮김. test 코드 리팩토링 진행. 최적화 진행

+++ flask_app/blog/auth.py
@@ -1,73 +1,28 @@
-import secrets
+from secrets import token_hex
 from flask import Blueprint, jsonify, redirect, url_for
 from flask_login import login_required, login_user, logout_user, current_user
+
 from .utils import (
-    logout_required, only_post_method,
-    not_have_create_permission_required, render_template_with_user,     # 데코레이터 함수
+    admin_required, logout_required, only_post_method,
+    not_have_create_permission_required, render_template_with_user,                 # 데코레이터 함수
 
-    form_invalid, success_msg, error_msg, get_method,                   # 기타
+    form_valid, form_invalid, success_msg, error_msg, get_method, post_method,      # 기타
 
-    smtp_setup, smtp_send_mail,
-    save_session, session_update, delete_session,
-    get_otp, get_remain_time,                                           # email 인증
+    smtp_send_mail, session_update, delete_session, get_otp, get_remain_time,       # email 인증
 
-    make_auth_url, get_access_token, get_user_info                      # third-party 인증
+    make_auth_url, get_access_token, get_user_info                                  # third-party 인증
 )
-from .forms import LoginForm, OtpForm, SignUpForm
+from .forms import CategoryForm, LoginForm, OtpForm, SignUpForm
 from .models import get_model
 
 auth = Blueprint('auth', __name__)
 BASE_AUTH_DIR = 'auth/'
 
 def render_template_auth(template_name_or_list, **context):
-    context['template_name_or_list'] = template_name_or_list
-    return render_template_with_user(BASE_AUTH_DIR, **context)
+    context['template_name_or_list'] = BASE_AUTH_DIR + template_name_or_list
+    return render_template_with_user(**context)
 
-@auth.route('/mypage', methods=['GET', 'POST'])
-@login_required
-def mypage():
-    session_update()
-    form = OtpForm()
-    session_otp = get_otp()
-    remain_time = get_remain_time()
-    params = {
-        'otp':session_otp,
-        'remain_time':remain_time,
-        'form':form,
-    }
-    
-    if get_method(): return render_template_auth("mypage.html", **params)
-
-    if form_invalid(form):
-        error_msg('otp 비밀번호는 6자리입니다. 길이를 맞춰주세요')
-        return render_template_auth("mypage.html", **params)
-    
-    otp = form.otp.data
-    if otp != session_otp:
-        error_msg('otp 비밀번호가 일치하지 않습니다.')
-        return render_template_auth("mypage.html", **params)
-    
-    delete_session()
-    current_user.save_instance(create_permission=True)
-    success_msg('이메일 인증 완료')
-    return render_template_auth("mypage.html")
-
-@auth.route('/send_mail_otp', methods=['POST'])
-@only_post_method
-@not_have_create_permission_required
-def send_mail_otp():
-    try:
-        smtp = smtp_setup()
-        otp = smtp_send_mail(smtp)
-        success_msg('인증번호 전송 완료')
-        save_session(otp)
-    except Exception as e:
-        error_msg(str(e))
-    finally:
-        smtp.quit()
-
-    return redirect(url_for('auth.mypage'))
-
+# ------------------------------------------------ 로그인, 로그아웃, 회원가입, 회원탈퇴 ------------------------------------------------
 @auth.route('/login', methods=['GET', 'POST'])
 @logout_required
 def login():
@@ -78,15 +33,10 @@ def login():
     }
 
     if get_method(): return render_template_auth('auth.html', **params)
+    if form_invalid(form): return render_template_auth('auth.html', **params)
     
-    if form_invalid(form):
-        error_msg('로그인 실패. 유효성 검사 실패')
-        return render_template_auth('auth.html', **params)
-    
-    user, status = get_model('user').user_check(email=form.email.data, password=form.password.data)
-    if not user:
-        error_msg(f'{status}')
-        return render_template_auth('auth.html', **params)
+    user = get_model('user').user_check(email=form.email.data, password=form.password.data)
+    if not user: return render_template_auth('auth.html', **params)
     
     login_user(user, remember=True)
     success_msg('로그인 성공!')
@@ -109,13 +59,8 @@ def sign_up():
     }
 
     if get_method(): return render_template_auth('auth.html', **params)
-    
-    if form_invalid(form):
-        error_msg('회원가입 실패. 유효성 검사 실패')
-        return render_template_auth('auth.html', **params)
-
+    if form_invalid(form): return render_template_auth('auth.html', **params)
     if get_model('user').duplicate_check(email=form.email.data, username=form.username.data):
-        error_msg('중복된 정보가 존재합니다')
         return render_template_auth('auth.html', **params)
     
     get_model('user')( 
@@ -131,12 +76,65 @@ def sign_up():
 @login_required
 def user_delete():
     user = get_model('user').get_instance_by_id(current_user.id)
+    if not user: return jsonify(message='error'), 404
     logout_user()
     
     user.delete_instance()
     success_msg('성공적으로 탈퇴하였습니다.')
     return jsonify(message='success'), 200
 
+# ------------------------------------------------ mypage, 이메일 인증 ------------------------------------------------
+@auth.route('/mypage', methods=['GET', 'POST'])
+@login_required
+def mypage():
+    session_update()
+    form = OtpForm()
+    session_otp = get_otp()
+    remain_time = get_remain_time()
+    params = {
+        'otp':session_otp,
+        'remain_time':remain_time,
+        'form':form,
+    }
+    
+    if get_method(): return render_template_auth("mypage.html", **params)
+    if form_invalid(form): return render_template_auth("mypage.html", **params)
+
+    if form.otp.data != session_otp:
+        error_msg('otp 비밀번호가 일치하지 않습니다.')
+        return render_template_auth("mypage.html", **params)
+    
+    delete_session()
+    current_user.save_instance(create_permission=True)
+    success_msg('이메일 인증 완료')
+    return render_template_auth("mypage.html")
+
+@auth.route('/send-mail-otp', methods=['POST'])
+@only_post_method
+@not_have_create_permission_required
+def send_mail_otp():
+    try:
+        smtp_send_mail()
+        success_msg('인증번호 전송 완료')
+    except Exception as e:
+        error_msg(str(e))
+
+    return redirect(url_for('auth.mypage'))
+
+@auth.route("/make-category", methods=['GET', 'POST'])
+@admin_required
+def make_category():
+    form = CategoryForm()
+
+    if post_method() and form_valid(form):
+        if not get_model('category').duplicate_check(name=form.name.data):
+            get_model('category')(name=form.name.data).add_instance()
+            success_msg(f'{form.name.data} category 생성에 성공하였습니다.')
+        form.name.data = ''
+
+    return render_template_auth("make_category.html", form=form)
+
+# ------------------------------------------------ third-party 가입 ------------------------------------------------
 @auth.route('/google/<string:type>')
 @logout_required
 def google_auth_page(type):
@@ -162,9 +160,7 @@ def callback_google(type):
             2. login_user(user)
         '''
         user = get_model('user').get_instance(email=response.json()['email'])
-        if not user:
-            error_msg('회원 가입이 필요한 이메일입니다.')
-            return redirect(url_for('views.home'))
+        if not user: return redirect(url_for('views.home'))
         
         login_user(user, remember=True)
         success_msg('로그인 성공!')
@@ -177,15 +173,13 @@ def callback_google(type):
             3. is_third_party = True로 계정 생성. 비밀번호는 랜덤 문자열 넣기 
         '''
         user_data = response.json()
-        user = get_model('user').duplicate_check(email=user_data['email'], username=user_data['name'])
-        if user:
-            error_msg('중복된 정보가 존재합니다.')
+        if get_model('user').duplicate_check(email=user_data['email'], username=user_data['name']): 
             return redirect(url_for('views.home'))
 
         get_model('user')(
             username=user_data['name'],
             email=user_data['email'],
-            password=secrets.token_hex(16),
+            password=token_hex(16),
             create_permission=True,
             is_third_party=True,
         ).add_instance()
